Move the test to the correct file.
