+++ metadata-ingestion/tests/unit/test_utilities.py
@@ -363,3 +363,22 @@ year(order_date)"""
     table_list = SqlLineageSQLParser(sql_query).get_tables()
     table_list.sort()
     assert table_list == ["order_items", "orders", "staffs"]
+
+
+def test_hash_in_sql_query_with_no_space():
+    parser = SqlLineageSQLParser(
+        sql_query="""
+/*
+HERE IS A STANDARD COMMENT BLOCK
+THIS WILL NOT BREAK sqllineage
+*/
+CREATE OR REPLACE TABLE `foo.bar.trg_tbl`AS
+#This, comment will not break sqllineage
+SELECT foo
+-- this comment will not break sqllineage either
+# this comment will not break sqllineage either
+FROM `foo.bar.src_tbl`
+        """
+    )
+
+    assert parser.get_tables() == ["foo.bar.src_tbl"]
